requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-dotenv
pydantic
httpx
aiofiles
python-multipart
re